Baroda-Finathon-Challenge/BarodaFinathon/requirements.txt
Django==1.11
djangorestframework
googlemaps
numpy
scipy
sklearn

